Open source ecosystems (#37)
This was a regular point within the FWD50 discussions.  I'd like to see more focus on building a vibrant ecosystem of collaborators than I am interested in the focus on the competitions.  Sure, competition can be useful, but investment in open communities from government is still very new in Canada.