Changing market to community
Making this change to use more open source jargon. Market support isn't as important as community support, and in any case, when companies contribute to an open source project they become part of the community. As such, I suggest community is a more apt term to use in this context.