Adding next step/Adjusting phrasing  (#43)
* Adding content to Development of guidelines header

Outlining processes required for ensuring requirements from the Official Languages Act is met.

* Changing market to community

Making this change to use more open source jargon. Market support isn't as important as community support, and in any case, when companies contribute to an open source project they become part of the community. As such, I suggest community is a more apt term to use in this context.

* Less aggressive wording

I perceived the previous phrasing to be a little aggressive. The change is intended to make it sound more formal/professional.