Free For Canadians
Proposing we take the approach that code produced by the Canadian government belongs to Canadians. The code would be made freely available for the all, however; they should not be able to profit from using said code base. The aim would be to assist companies within Canada giving them code to work with with the aim of assisting with the development of a Canadian technological ecosystem to support and create new projects.