Phrasing + Benefit for Canadians (#50)
* Free For Canadians 

Proposing we take the approach that code produced by the Canadian government belongs to Canadians. The code would be made freely available for the all, however; they should not be able to profit from using said code base. The aim would be to assist companies within Canada giving them code to work with with the aim of assisting with the development of a Canadian technological ecosystem to support and create new projects.

* HTML5 and CSS3

We could also just generically add HTML*, CSS* ...
Figured if we are going to specify HTML5 we may as well be consistent and specify CSS3 as well.

* Redundant wording

**Interoperability** 'can be defined as' an ... 

It can be defined as it given that we're providing a definition here.

Acknowledging the nuances and exceptions put forth by other areas of this white paper (especially those in the source code section) wording should be added in to allow for flexibility as requirements may vary by projects.

* Remove default to reciprocal licences