OS Web Dev Standard Recommendations  (#59)
* Recommendations for database web dev software 

Added 'particularly for primary functions for business critical applications' because for certain functions in applications, or smaller applications it is not as important to abide by standards. The smaller the code base the easier to adjust if need be if say, for example, a code base declines in popularity. For example, mongodb or Redis (or some other NoSQL database) could be used for a chat function on an app (fast read/write) but a core business application are much more likely to rely on SQL / Relational DB principles)

* Recommendations - web server 

Adding in recommendations for java web servers. Preference for Tomcat or if application server is a requirement than Wildfly followed by GlassFish.